+++ requre/__init__.py
@@ -1,8 +1,13 @@
-from requre.import_system import decorate, replace
+# Copyright Contributors to the Packit project.
+# SPDX-License-Identifier: MIT
+
 from requre.base_testclass import RequreTestCase
+from requre.import_system import decorate, replace
+from requre.record_and_replace import record
 
 __all__ = [
     decorate.__name__,
     replace.__name__,
+    record.__name__,
     RequreTestCase.__name__,
 ]
