+++ requre/import_system.py
@@ -1,39 +1,19 @@
-# MIT License
-#
-# Copyright (c) 2018-2019 Red Hat, Inc.
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
+# Copyright Contributors to the Packit project.
+# SPDX-License-Identifier: MIT
 
 import builtins
 import functools
 from enum import Enum
-from requre.constants import DEFAULT_IMPORT_FUNCTION
-from typing import Optional, Any, Callable, List
+from typing import Any, Callable, List, Optional
 
-from requre.storage import PersistentObjectStorage
 from requre.cassette import Cassette
-from requre.online_replacing import (
+from requre.constants import DEFAULT_IMPORT_FUNCTION
+from requre.record_and_replace import (
     ModuleRecord,
     _parse_and_replace_sys_modules,
     _revert_modules,
 )
+from requre.storage import PersistentObjectStorage
 
 
 class ReplaceType(Enum):
