+++ tests/data/e2e_latency_replacements.py
@@ -1,4 +1,4 @@
 from requre.import_system import UpgradeImportSystem
-from requre.helpers.simple_object import Simple
+from requre.simple_object import Simple
 
 FILTERS = UpgradeImportSystem().decorate("time.sleep", Simple.decorator_plain())
